requirements.txt
click
pyyaml
tensorboard
tqdm
lpips
piq
torchdiffeq